Refactor code and remove Async SQS
This removes the async SQS and swaps it out for synchronous message
sends instead. It also refactors a few functions to account for this
change.

The ability to send both batched and single messages is kept, but it
might make more sense to go down to just single messages as we're not
really needing to send lots of individual messages.

I'll commit both methods here, for psoterity, and then remove batch
messages in a future commit.