Remove batched messages
Removes the functionality to send batched messages. These aren't
necessary when we can build a single larger message of the same size
instead and reduce complexity in the process.

Also adds in a specific conversion method for the CleanseList to Json
step to make sure that the correct Type is parsed into a string before
sending upstream.