Query for getting active user list length
To properly assess the impact of the list cleanse and to make sure that
we can implement a circuitbreaker to kick in if the cleanse fraction
looks too high, we need to be able to query for the number of current
active users.

This adds that query in here and implements it in the `getCleanseList`
lambda.