fix over-selection for users of Today UK
This fixes an issue observed in production where we over-selected (and therefore over-unsubscribed) users.
This was due to the fact the request would select users who read the weekdays *and* the weekend version of the newsletter over the last 3 months, where we wanted the weekdays *or* the weekend version of the newsletter.

The fix is pretty simple, don't join on the campaign name which will restrict which event we'll filter on, but re-inject the two campaign names in the request.