Make the BrazeData handling simpler
Rather than using a fancy partially applied function, I just explicitly
add the BrazeData as part of a new case class. This should be easier to
read and debug in future.

Getting the Braze data for a given newsletter is done using `find`
which I think should also be easier to debug.